Refactor tests to use resources package
