+++ olio/service/resources/version_resource_test.go
@@ -1,4 +1,4 @@
-package resources_test
+package resources
 
 import (
 	"net/http"
@@ -6,27 +6,26 @@ import (
 	"testing"
 
 	"github.com/gin-gonic/gin"
-	"github.com/olioapps/service-skeleton-go/olio/service/resources"
 	"github.com/stretchr/testify/assert"
 )
 
-type VersionExtractor struct {
+type Extractor struct {
 }
 
-func NewVersionExtractor() *VersionExtractor {
-	return &VersionExtractor{}
+func NewExtractor() Extractor {
+	return Extractor{}
 }
 
-func (ve VersionExtractor) GetVersion() string {
+func (e Extractor) GetVersion() string {
 	return "0.0.11"
 }
 
-func (ve VersionExtractor) GetAppName() string {
+func (e Extractor) GetAppName() string {
 	return "cx-messaging"
 }
 
 func TestVersion(t *testing.T) {
-	versionExtractor := NewVersionExtractor()
+	versionExtractor := NewExtractor()
 
 	tt := []struct {
 		name string
@@ -39,23 +38,23 @@ func TestVersion(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			if tc.name == "No extractor" {
 				router := gin.New()
-				versionResource := resources.NewVersionResource()
-				router.GET("/api/version", versionResource.GetVersion)
+				versionResource := NewVersionResource()
+				router.GET("/api/version", versionResource.getVersion)
 				req, _ := http.NewRequest("GET", "/api/version", nil)
 				res := httptest.NewRecorder()
 				router.ServeHTTP(res, req)
-				assert.Equal(t, "{\n    \"serviceFrameworkVersion\": \""+resources.VERSION+"\",\n    \"appVersion\": \"no version given\"\n}", res.Body.String())
+				assert.Equal(t, "{\n    \"serviceFrameworkVersion\": \""+VERSION+"\",\n    \"appVersion\": \"no version given\"\n}", res.Body.String())
 			}
 
 			if tc.name == "With extractor" {
 				router := gin.New()
-				versionResource := resources.NewVersionResource()
+				versionResource := NewVersionResource()
 				versionResource.AddVersionExtractor(versionExtractor)
-				router.GET("/api/version", versionResource.GetVersion)
+				router.GET("/api/version", versionResource.getVersion)
 				req, _ := http.NewRequest("GET", "/api/version", nil)
 				res := httptest.NewRecorder()
 				router.ServeHTTP(res, req)
-				assert.Equal(t, "{\n    \"serviceFrameworkVersion\": \""+resources.VERSION+"\",\n    \"appVersion\": \"cx-messaging-0.0.11\"\n}", res.Body.String())
+				assert.Equal(t, "{\n    \"serviceFrameworkVersion\": \""+VERSION+"\",\n    \"appVersion\": \"cx-messaging-0.0.11\"\n}", res.Body.String())
 			}
 		})
 	}
